+++ src/chains/supported.chains.ts
@@ -952,7 +952,7 @@ export const supportedEVMChains: EVMChain[] = [
         symbol: 'ETH',
         decimals: 18,
       },
-      rpcUrls: ['https://goerli.infura.io/v3/d1caeba320f94122ba8f791f50122c4c'],
+      rpcUrls: ['https://rpc.ankr.com/eth_goerli'],
     },
   },
   // 12 - Metadium Testnet
@@ -1097,7 +1097,7 @@ export const supportedEVMChains: EVMChain[] = [
 
     metamask: {
       chainId: prefixChainId(421613),
-      blockExplorerUrls: ['https://goerli-rollup-explorer.arbitrum.io'],
+      blockExplorerUrls: ['https://goerli-rollup-explorer.arbitrum.io/'],
       chainName: 'Arbitrum Görli',
       nativeCurrency: {
         name: 'AGOR',
@@ -1203,10 +1203,7 @@ export const supportedEVMChains: EVMChain[] = [
         symbol: 'HT',
         decimals: 18,
       },
-      rpcUrls: [
-        'https://http-testnet.hecochain.com',
-        'wss://ws-testnet.hecochain.com',
-      ],
+      rpcUrls: ['https://http-testnet.hecochain.com'],
     },
   },
   // 1666700000 - Harmony Testnet 0
@@ -1324,7 +1321,7 @@ export const supportedEVMChains: EVMChain[] = [
     // https://github.com/ethereum-lists/chains/blob/master/_data/chains/eip155-43113.json
     metamask: {
       chainId: prefixChainId(43113),
-      blockExplorerUrls: ['https://testnet.snowtrace.io'],
+      blockExplorerUrls: ['https://testnet.snowtrace.io/'],
       chainName: 'Avalanche Fuji Testnet',
       nativeCurrency: {
         name: 'AVAX',
@@ -1349,7 +1346,7 @@ export const supportedEVMChains: EVMChain[] = [
 
     metamask: {
       chainId: prefixChainId(59140),
-      blockExplorerUrls: ['https://explorer.goerli.zkevm.consensys.net'],
+      blockExplorerUrls: ['https://explorer.goerli.zkevm.consensys.net/'],
       chainName: 'ConsenSys zkEVM Testnet',
       nativeCurrency: {
         name: 'crETH',